move db specific options to BlockchainDB
Avoids common depending on blockchain_db, which can cause
link errors.
